# Improving speed by using `eq_ignore_ascii_case()` before `regular expression match`