I found that Rust's String struct has a function for `case-insensitive matching` called [eq_ignore_ascii_case()](https://doc.rust-lang.org/std/string/struct.String.html#method.eq_ignore_ascii_case) :)
Since  this function allocates no memory and fast, I added [eq_ignore_ascii_case()](https://doc.rust-lang.org/std/string/struct.String.html#method.eq_ignore_ascii_case) match.

## What Changed
Added `eq_ignore_ascii_case()` before `regular expression match` to improve speed.
 - Check with [eq_ignore_ascii_case()](https://doc.rust-lang.org/std/string/struct.String.html#method.eq_ignore_ascii_case) only when [string length matched](https://github.com/Yamato-Security/hayabusa/pull/883).
   - This reduces the number of `regular expression matches` and makes it faster 

## Evidence

### Environment
- OS: Windows 10 Home edition
- Hard: Memory 16GB , Core 8, SSD, laptop

### Benchmark1
I ran a benchmark using [this procedure(6.1GB evtx)](https://github.com/Yamato-Security/hayabusa/issues/778#issuecomment-1296504766) and the results were as follows.

|Version| Elapsed time | Memory(peak)  | Events with hits / Total events | Output file size(bytes)|
|-------| ------------ | --------|--------|--------|
| [before](https://github.com/Yamato-Security/hayabusa/commit/f13f0bb968c934fa00d07aad030abe41d1297063)  | 00:12:08.121| 5.0 GiB| 1,593,715 / 4,817,181 | 575085389 | 
| [This PR](https://github.com/Yamato-Security/hayabusa/commit/e34530722823a7c650561617488ae540b4c0c64d) | 00:11:38.499| 4.8 GiB| 1,593,715 / 4,817,181 | 575085389 | 

#### Console output
[before](https://github.com/Yamato-Security/hayabusa/commit/31cd18e0bf9d13e67e9cf3a262d80af39bb980b2)
```
PS C:\tmp\hayabusa-2.1.0-win-64-bit> .\hayabusa.exe csv-timeline -d ..\hayabusa-big-evtx\ -o 1.csv --debug
Results Summary:

Events with hits / Total events: 1,593,715 / 4,817,181 (Data reduction: 3,223,466 events (66.92%))

Total | Unique detections: 1,627,284 | 150
Total | Unique critical detections: 0 (0.00%) | 0 (0.00%)
Total | Unique high detections: 12,044 (0.74%) | 20 (13.33%)
Total | Unique medium detections: 11,118 (0.68%) | 38 (25.33%)
Total | Unique low detections: 1,053,623 (64.75%) | 42 (28.00%)
Total | Unique informational detections: 550,499 (33.83%) | 50 (33.33%)
...
Saved file: 1.csv (575.1 MB)
Elapsed time: 00:12:08.121
Rule Parse Processing Time: 00:00:20.157
Analysis Processing Time: 00:11:26.545
Output Processing Time: 00:00:21.417

Memory usage stats:
heap stats:    peak      total      freed    current       unit      count
  reserved:    5.0 GiB    5.0 GiB   83.0 MiB    4.9 GiB
 committed:    4.6 GiB   56.7 GiB   52.2 GiB    4.5 GiB
```

[This PR](https://github.com/Yamato-Security/hayabusa/commit/f13f0bb968c934fa00d07aad030abe41d1297063)
```
PS C:\tmp\hayabusa-2.1.0-win-64-bit> .\hayabusa.exe csv-timeline -d ..\hayabusa-big-evtx\ -o 1.csv --debug
Results Summary:

Events with hits / Total events: 1,593,715 / 4,817,181 (Data reduction: 3,223,466 events (66.92%))

Total | Unique detections: 1,627,284 | 150
Total | Unique critical detections: 0 (0.00%) | 0 (0.00%)
Total | Unique high detections: 12,044 (0.74%) | 20 (13.33%)
Total | Unique medium detections: 11,118 (0.68%) | 38 (25.33%)
Total | Unique low detections: 1,053,623 (64.75%) | 42 (28.00%)
Total | Unique informational detections: 550,499 (33.83%) | 50 (33.33%)
...
Saved file: 1.csv (575.1 MB)
Elapsed time: 00:11:38.499
Rule Parse Processing Time: 00:00:21.788
Analysis Processing Time: 00:10:55.151
Output Processing Time: 00:00:21.559

Memory usage stats:
heap stats:    peak      total      freed    current       unit      count
  reserved:    4.8 GiB    4.8 GiB   83.0 MiB    4.7 GiB
 committed:    4.4 GiB   54.0 GiB   49.7 GiB    4.2 GiB
```

### Benchmark2
I ran a benchmark using [hayabusa-sample-evtx](https://github.com/Yamato-Security/hayabusa-sample-evtx) and the results were as follows.

|Version| Elapsed time | Memory(peak)  | Events with hits / Total events | Output file size(bytes)|
|-------| ------------ | --------|--------|--------|
| [before](https://github.com/Yamato-Security/hayabusa/commit/f13f0bb968c934fa00d07aad030abe41d1297063)  | 00:00:17.135| 2.1 GiB| 19,606 / 47,458 | 16448088 | 
| [This PR](https://github.com/Yamato-Security/hayabusa/commit/e34530722823a7c650561617488ae540b4c0c64d) | 00:00:16.653| 1.9 GiB| 19,606 / 47,458 | 16448088 | 

I would appreciate it if you could review🙏