@hitenkoku 2023-01-18
@YamatoSecurity Source code reviewed. I do not think there is a problem.

@fukusuket Thank you so much for your pull request which will greatly improve performance.