@fukusuket 2023-01-18
Thank you so much for benchmark and code review !
It was surprising that it was effective in reducing memory usage😲  I will continue to look for other places where regex can be reduced.